Update "Right Panel" to use SlidePanel from Listing Page
