+++ src/Panel/ExpandableSection.jsx
